# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LendSwift Loan Application API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ApplicationLoanType = typeof ApplicationLoanType[keyof typeof ApplicationLoanType];


export const ApplicationLoanType = {
  personal: 'personal',
  home: 'home',
  business: 'business',
} as const;

export type ApplicationStatus = typeof ApplicationStatus[keyof typeof ApplicationStatus];


export const ApplicationStatus = {
  draft: 'draft',
  submitted: 'submitted',
  approved: 'approved',
  rejected: 'rejected',
} as const;

/**
 * @nullable
 */
export type ApplicationFormData = { [key: string]: unknown } | null;

export interface Application {
  id: string;
  loanType: ApplicationLoanType;
  currentStep: number;
  status: ApplicationStatus;
  /** @nullable */
  formData?: ApplicationFormData;
  createdAt: string;
  updatedAt: string;
}

export type ApplicationInputLoanType = typeof ApplicationInputLoanType[keyof typeof ApplicationInputLoanType];


export const ApplicationInputLoanType = {
  personal: 'personal',
  home: 'home',
  business: 'business',
} as const;

/**
 * @nullable
 */
export type ApplicationInputFormData = { [key: string]: unknown } | null;

export interface ApplicationInput {
  loanType: ApplicationInputLoanType;
  /** @nullable */
  formData?: ApplicationInputFormData;
  /** @nullable */
  currentStep?: number | null;
}

/**
 * @nullable
 */
export type ApplicationUpdateFormData = { [key: string]: unknown } | null;

export interface ApplicationUpdate {
  /** @nullable */
  currentStep?: number | null;
  /** @nullable */
  formData?: ApplicationUpdateFormData;
}

export interface SubmitResult {
  id: string;
  referenceNumber: string;
  status: string;
  message: string;
}

export type PreApprovalSummaryVerdict = typeof PreApprovalSummaryVerdict[keyof typeof PreApprovalSummaryVerdict];


export const PreApprovalSummaryVerdict = {
  likely_approved: 'likely_approved',
  review_required: 'review_required',
  likely_rejected: 'likely_rejected',
} as const;

export interface PreApprovalSummary {
  applicationId: string;
  loanType: string;
  eligibilityScore: number;
  indicativeRate: number;
  maxAmount: number;
  verdict: PreApprovalSummaryVerdict;
  reasons?: string[];
  /** @nullable */
  applicantName?: string | null;
  /** @nullable */
  requestedAmount?: number | null;
  /** @nullable */
  tenure?: number | null;
}

export interface PanVerifyInput {
  pan: string;
  /** @nullable */
  name?: string | null;
}

export interface AadhaarVerifyInput {
  aadhaar: string;
  /** @nullable */
  otp?: string | null;
}

/**
 * @nullable
 */
export type VerificationResultDetails = { [key: string]: unknown } | null;

export interface VerificationResult {
  verified: boolean;
  message: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  details?: VerificationResultDetails;
}

export interface AddressSuggestion {
  label: string;
  city: string;
  state: string;
  pincode: string;
  /** @nullable */
  district?: string | null;
}

export interface DocumentInput {
  docType: string;
  fileName: string;
  fileSize: number;
  /** @nullable */
  mimeType?: string | null;
  /** @nullable */
  base64Preview?: string | null;
}

export interface Document {
  id: string;
  applicationId: string;
  docType: string;
  fileName: string;
  fileSize: number;
  /** @nullable */
  mimeType?: string | null;
  createdAt: string;
}

export type AutocompleteAddressParams = {
query: string;
pincode?: string;
};