# lib/api-zod/src/generated/types/addressSuggestion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LendSwift Loan Application API
 * OpenAPI spec version: 0.1.0
 */

export interface AddressSuggestion {
  label: string;
  city: string;
  state: string;
  pincode: string;
  /** @nullable */
  district?: string | null;
}