# lib/api-zod/src/generated/types/applicationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LendSwift Loan Application API
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationInputFormData } from './applicationInputFormData';
import type { ApplicationInputLoanType } from './applicationInputLoanType';

export interface ApplicationInput {
  loanType: ApplicationInputLoanType;
  /** @nullable */
  formData?: ApplicationInputFormData;
  /** @nullable */
  currentStep?: number | null;
}