# lib/api-zod/src/generated/types/applicationUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LendSwift Loan Application API
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationUpdateFormData } from './applicationUpdateFormData';

export interface ApplicationUpdate {
  /** @nullable */
  currentStep?: number | null;
  /** @nullable */
  formData?: ApplicationUpdateFormData;
}