# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LendSwift Loan Application API
 * OpenAPI spec version: 0.1.0
 */

export interface Document {
  id: string;
  applicationId: string;
  docType: string;
  fileName: string;
  fileSize: number;
  /** @nullable */
  mimeType?: string | null;
  createdAt: string;
}