# lib/api-zod/src/generated/types/panVerifyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LendSwift Loan Application API
 * OpenAPI spec version: 0.1.0
 */

export interface PanVerifyInput {
  pan: string;
  /** @nullable */
  name?: string | null;
}