# lib/api-zod/src/generated/types/submitResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LendSwift Loan Application API
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitResult {
  id: string;
  referenceNumber: string;
  status: string;
  message: string;
}