# lib/api-zod/src/generated/types/verificationResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LendSwift Loan Application API
 * OpenAPI spec version: 0.1.0
 */
import type { VerificationResultDetails } from './verificationResultDetails';

export interface VerificationResult {
  verified: boolean;
  message: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  details?: VerificationResultDetails;
}